Add files via upload
was screwing around with this code. Bright side: it does technically work and does use binary heap. downsides: 1) i realized the hueristic is so basic that it doesn't actually take into account any blocked squares (so will occasionally make a dumb move because the heuristic wants to take 1 step away from the target to magically go through the blocked spaces to take a more direct route.) It also can probably be coded slightly better. Finally, I've partially fixed this, but I dont have a specific tiebreak for if estimated cost is the same, so it can be stupid about catcorners ( ie when square of 4 with target on one corner and  start opposite corner, it has a tendency to first go one way, then stop and go the other way (beacuse g+h is 2 for both), and only then get to the target.